+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Most recent releases are shown at the top. Each release shows:
 - Add tests for dynamic field storage and retrieval (#190)
 - Simplify vectorstore schemas by removing fields that can be dynamically added (#200)
 - change `persist_directory` to `persist_location` in stores (#200)
+- make content, id, and source fields dynamic to support existing ES indices (#202)
 
 ### fixed:
 - code improvements/fixes thanks to @Mirza-Samad-Ahmed-Baig (#188)